requirements.txt
deepface==0.0.93
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
opencv-python==4.11.0.86
tensorflow==2.19.0
uvicorn==0.34.2
mediapipe
tf-keras
dotenv
